+++ wallet/network/store/v1/store.go
@@ -1,6 +1,7 @@
 package v1
 
 import (
+	"errors"
 	"fmt"
 	"io/fs"
 	"os"
@@ -15,6 +16,12 @@ import (
 )
 
 const fileExt = ".toml"
+
+var (
+	ErrNetworkNameCannotBeEmpty      = errors.New("the network name cannot be empty")
+	ErrNetworkNameCannotStartWithDot = errors.New("the network name cannot start with a `.`")
+	ErrNetworkNameCannotContainSlash = errors.New("the network name cannot contain `\\` nor `/`")
+)
 
 type Store struct {
 	networksHome string
@@ -43,6 +50,10 @@ func (s *Store) ListNetworks() ([]string, error) {
 	}
 	networks := []string{}
 	for _, entry := range entries {
+		if err := ensureValidNetworkName(entry.Name()); err != nil {
+			continue
+		}
+
 		if strings.HasSuffix(entry.Name(), ".toml") {
 			networks = append(networks, s.fileNameToName(entry.Name()))
 		}
@@ -60,10 +71,18 @@ func (s *Store) GetNetworkPath(name string) string {
 }
 
 func (s *Store) NetworkExists(name string) (bool, error) {
+	if err := ensureValidNetworkName(name); err != nil {
+		return false, err
+	}
+
 	return vgfs.FileExists(s.GetNetworkPath(name))
 }
 
 func (s *Store) GetNetwork(name string) (*network.Network, error) {
+	if err := ensureValidNetworkName(name); err != nil {
+		return nil, err
+	}
+
 	nfc := &networkFileContent{}
 	if err := paths.ReadStructuredFile(s.nameToFilePath(name), &nfc); err != nil {
 		return nil, fmt.Errorf("couldn't read network configuration file: %w", err)
@@ -75,6 +94,10 @@ func (s *Store) GetNetwork(name string) (*network.Network, error) {
 }
 
 func (s *Store) SaveNetwork(net *network.Network) error {
+	if err := ensureValidNetworkName(net.Name); err != nil {
+		return err
+	}
+
 	nfc := &networkFileContent{
 		API: net.API,
 	}
@@ -85,6 +108,9 @@ func (s *Store) SaveNetwork(net *network.Network) error {
 }
 
 func (s *Store) DeleteNetwork(name string) error {
+	if err := ensureValidNetworkName(name); err != nil {
+		return err
+	}
 	path := s.GetNetworkPath(name)
 	return os.Remove(path)
 }
@@ -95,4 +121,19 @@ func (s *Store) nameToFilePath(network string) string {
 
 func (s *Store) fileNameToName(fileName string) string {
 	return fileName[:len(fileName)-len(fileExt)]
+}
+
+func ensureValidNetworkName(name string) error {
+	if name == "" {
+		return ErrNetworkNameCannotBeEmpty
+	}
+	if name[0] == '.' {
+		return ErrNetworkNameCannotStartWithDot
+	}
+
+	if strings.ContainsAny(name, "\\/") {
+		return ErrNetworkNameCannotContainSlash
+	}
+
+	return nil
 }